Move authentication-related functions from wagtail.admin.utils to wagtail.admin.auth
